refactor(assistant): move onboarding-events-store out of memory/ into onboarding/
