+++ assistant/src/__tests__/conversation-surfaces-activation-emit.test.ts
@@ -39,7 +39,7 @@ import {
   isActivationSession,
   markActivationSession,
 } from "../memory/activation-session-store.js";
-import { queryUnreportedOnboardingEvents } from "../memory/onboarding-events-store.js";
+import { queryUnreportedOnboardingEvents } from "../onboarding/onboarding-events-store.js";
 import { getDb } from "../persistence/db-connection.js";
 import { initializeDb } from "../persistence/db-init.js";
 import {
