+++ assistant/src/onboarding/__tests__/onboarding-events-store.test.ts
