+++ assistant/src/onboarding/onboarding-events-store.ts
